requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.3
argcomplete==3.6.0
certifi==2025.1.31
chardet==3.0.4
click==7.1.1
coloredlogs==14.0
decorator==4.4.2
distlib==0.3.6
distro==1.9.0
docutils==0.15.2
dogpile.cache==0.9.0
dynaconf==2.2.3
filelock==3.9.0
gitdb==4.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==8.1
idna==2.9
iso8601==0.1.12
Jinja2==2.11.1
jmespath==0.9.5
jsonpatch==1.25
jsonpointer==2.0
llama_stack_client==0.1.6
markdown-it-py==3.0.0
MarkupSafe==1.1.1
mdurl==0.1.2
munch==2.5.0
netifaces==0.10.9
nvidia-ml-py==12.570.86
os-service-types==1.7.0
pandas==2.2.3
pbr==5.4.5
pipenv==2022.12.19
platformdirs==2.6.2
prompt_toolkit==3.0.50
pyaml==25.1.0
pycparser==2.20
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynvml==12.0.0
python-box==3.4.6
python-dateutil==2.9.0.post0
python-dotenv==0.12.0
pytz==2025.1
PyYAML==5.3.1
ramalama==0.6.3
rank-bm25==0.2.2
requests==2.23.0
requestsexceptions==1.4.0
rich==13.9.4
six==1.14.0
smmap==3.0.1
sniffio==1.3.1
soupsieve==2.0
stevedore==1.32.0
termcolor==2.5.0
toml==0.10.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.25.8
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
